new feature added

+++ WebScraping.Core.Domain/Settings/JWTSettings.cs
@@ -2,7 +2,7 @@
 {
     public class JWTSettings
     {
-        public double DurationInMinutes { get; set; }
+        public int DurationInMinutes { get; set; }
         public string Audience { get; set; }
         public string Issuer { get; set; }
         public string Key { get; set; }
